requirements.txt
img2pdf
pyvips
requests
selenium
bs4
lxml
python-telegram-bot